Fixed RandomEventTrigger and ScreenShake.
RandomEventTrigger  and Screenshake has been added to the gamescene inside the GameManager.
GameManager has also gotten an AudioSource component with the train sound effect on it.

RandomEventTrigger:
function ResetMotivation has been fixed.
getcomponent<audiosource> no longer tries to pull it from the camera (should not exist an audiosource on the camera).
added a float for motivation duration, as using shakeDuration for how long the motivation was zero seemed wrong.
created a coroutine for starting the invokeRepeat for TriggerRandomEvent(). Now it is invoked when all NPCs has been created instead and starts 2 seconds after that.

ScreenShake:
removed an unused variable from RandomEventTrigger which screenShake tried to call and set to false, but still did nothing.
made smaller changes.