NPCs will now be "tracked" when they feel bad.
When an NPC has less than half of its total feelings (if happiness and motivation combined is only half the total possible value)
The script, a canvas and a temp arrow has been added to the Game Scene.
Some Arist Variables was created for the script aswell.

Can now get TotalFeelings from an NPC.