Updated radiator prefab. Added script.
Radiator prefabs now has a new script on them (on the canvas gameobject), this will rotate the button towards the camera at all times.

Script added "RotateObjectTowardsCamera" is a script which rotates the objects forward towards the cameras forward transform.
The script has a few inspector variables if needed to tweak the rotation and reversity of the object.