ToiletSystem at rough usable state.
NPCs now stand in line to get to the toilet.
The toilet queue is added as soon as an NPC is on it's way to the toilet (meaning its not first there, first serve. But whoever decides they need to go to the toilet first, first.).
Is messy and needs some cleaning. Very possible it's not performance friendly either (but should be somewhat fine).

Changed NPC collision box and agent avoidance box to much smaller than before, for an easier time of walking past other agents.