Started better radiators, did many different stuff.
Created a script to hold prefabs (to avoid creating new prefab holders in each script), these prefabs are put into a dictionary for possibly easier getting.
Got a script from the internetz to make particles visible in the UI. Created a new particle with this for feedback to showing radiators draining happiness.
NpcButtons is responsible for the particle.